+++ src/Api/DTO/PersonDto/GetUserDto.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Exadel.OfficeBooking.Api.DTO.PersonDto
+{
+    public class GetUserDto : SetUserDto
+    {
+        public Guid Id { get; set; }
+    }
+}
