+++ src/Api/DTO/PersonDto/GetVacationDto.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Exadel.OfficeBooking.Api.DTO.PersonDto
+{
+    public class GetVacationDto : SetVacationDto
+    {
+        public Guid Id { get; set; }
+    }
+}
