+++ src/Api/DTO/PersonDto/SetVacationDto.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+
+namespace Exadel.OfficeBooking.Api.DTO.PersonDto
+{
+    public class SetVacationDto
+    {
+        [Required]
+        public Guid UserId { get; set; }
+
+        [Required]
+        public DateTime VacationStart { get; set; }
+        
+        [Required]
+        public DateTime VacationEnd { get; set; }
+    }
+}
